libdumpalloc: Work-around invalid, mis-aligned ip when scanning beyond start of code section.
Even though we know where the starting virtual address of each *segment*,
this may contain other data apart from executable instructions.

If we stray into a data area when back-scanning (to reconstruct stack), it
is possible that we mis-interpret some data as a stack adjustment, just
because the upper 16 bits match the `addiu sp, sp` signature. This will
result in a bad stack pointer, and probably a segfault.

A proper solution is required for this, but for now this commit at least
catches the case where the resulting sp is mis-aligned. It won't help in
the case where the sp is wrong, but still aligned.