libdumpalloc: Capture allocation & deallocation times.
This commit adds a timestamp (in seconds) to each record written. This will
help with analysing memory usage over time (and hence in pinpointing
memory leaks).

The monotonic clock is used in preference to wall-clock when obtaining the
timestamps to avoid potential issues with the system clock being reset
whilst capture is taking place.