readalloc: Read timestamps from `libdumpalloc` records.
This commit updates `readalloc` such that it is able to parse the new
`libdumpalloc` record format that includes timestamps on each record.

It doesn't do anything useful with the timestamps yet. It just stores them
in the `Object`, and `Alloc` data structures. In the future though, the
timestamp information will enable reports to be periodically generated, and
different timeslices to be taken through the allocation data.