libdumpalloc: Dial-down optimisation to -O2 to work-around problem with MIPS backtracing.
When `libdumpalloc.so` is built with `-O3`, it was found that MIPS
backtracing would go badly wrong - even managing to calculate mis-aligned
(odd) return-addresses.

Right now, it is unclear exactly why/how this was happening, and the problem
requires further investigation. However, what is known is that it would
sometimes calculate an odd `ra`, which would then cause the backward-scanning
in the stack walker to walk right off the current segment - resulting in a
segfault.

Dialling-down the optimisation level appears to make the problem go away.