readalloc: Generate a report of unresolved symbols at end of run.
`readalloc` receives stack dumps from `libdumpalloc`, in the form of an
ELF object reference, plus a virtual address for each frame.

It tries to resolve each of the virtual addresses to a symbol within the
relevant ELF, but it fails to do this for some symbols. We would like to
know which symbols cannot be resolved, such that they can be manually
checked.

This commit modifies `readalloc` such that it keeps track of the addresses
it fails to resolve for each object, and generates a single report file at
the end of the run.