readalloc: Add time-sliced report generation.
Now that `libdumpalloc` is providing allocation timestamps, we can
split the active allocations according to when they were made and
generate a separate report for each time period.

This commit adds such a feature. It is activated via a new cmdline
option: `--time-slice-interval-secs`.

For example, to generate a separate report for allocations made
over each hourly period:

  cat dumpalloc-file | readalloc --time-slice-interval-secs 3600 /path/to/device/rootfs

Note: even when this new option is used, the generated reports will
still only include the allocations that were still active at the
end of the overall measurement period. The purpose of adding this
feature was to make it easier to spot memory leaks, and so we are
still only interested in the memory that is never freed.