libdumpalloc: Add support for dumping Python stack frames.
In cases where we have a Python runtime embedded in our program
we would like to know which Python functions were ultimately
responsible for causing memory to be allocated. A native stack
dump doesn't tell us a great deal in this case, since the
Python runtime will be allocating objects from pre-allocated
arenas.

This commit adds experimental support for gathering Python
stack frames, in addition to native frames.

At the moment this is quite crude, and not strictly accurate,
in the sense that the Python frames are presented after the
native frames (rather than intermixed as they should be).

We are also only capturing Python stack frames when some
object creation in Python triggers the runtime to allocate a
new arena (whilst it would be useful if we could track all
the individual Python allocations).

The Python support is built into a separate shared-library.
To enable it, users must run their executable with:

    LD_PRELOAD="libdumpalloc-python.so libdumpalloc.so" myprog

rather than just:

    LD_PRELOAD="libdumpalloc.so" myprog

NOTE:

As Python is a relatively-heavy dependency, the separate shared
object: `libdumpalloc-python.so` is only built if the Python
libraries are already available in the staging environment.