libdumpalloc: Initialise pointer to real calloc() function before others.
It looks like calling dlsym() on PC causes a call to calloc(). Since dlsym()
is used multiple times during init() to lookup the "real" allocation
functions, this leads to a segmentation fault as `real_calloc` is not yet
initialised.

This commit does 2 things:

1) Re-orders init() such that the real calloc() function is looked-up first.

2) Modifies the substitute calloc() implementation such that it returns a
NULL pointer if `real_calloc` is not yet initialised. This seems to be
enough to allow the first dlsym() to succeed.