libdumpalloc: Deal with dl_iterate_phdr() returning objects with zero-length names.
dl_iterate_phdr() is used to enumerate the loaded Elf objects (libraries
and the executable itself). The base address and name (full path) are then
dumped out as part of the generated output. This information can then later
be used by a tool such as `readalloc` to help resolve symbol addresses to
names.

When running on PC, against Glibc, not all objects returned by
dl_iterate_phdr() have a name. In particular, the executable itself, and
the "fake" shared-object: `linux-vdso.so` are missing names.

This commit fixes-up the names.