WIP: Interpose PyObject_Malloc() and PyObject_Free()
As far as I can tell, all object allocations in Python will result in a call
to PyObject_Malloc(). This is also the only function that calls `new_arena()`.

Also, it looks like all object deletions will result in a call to `PyObject_Free()`.

It is not so useful to capture only when `new_arena()` invokes `malloc()`,
since we only then see the Python operation that caused a new arena to be
created, not what cause all the existing ones to fill-up.